add example to toc, import and docs cleanup, linting
