+++ docs/examples.rst
@@ -8,6 +8,7 @@ This page groups links to example use cases of votelib.
    :caption: Examples:
 
    examples/sk_nr_2020.ipynb
+   examples/cz_psp_2017.ipynb
    examples/de_bdt_2017.ipynb
    examples/ie_pres_1990.ipynb
    examples/nmnmet_cc_2018.ipynb
