+++ votelib/util.py
@@ -13,7 +13,7 @@ from typing import Any, List, Tuple, Dict, FrozenSet, Union
 from numbers import Number
 
 from .vote import RankedVoteType
-from .candidate import Candidate, Constituency
+from .candidate import Candidate
 
 
 def add_dict_to_dict(dict1: Dict[Any, Number],
@@ -110,40 +110,6 @@ def _select_n_random_int(candidates: List[Any],
             + [wt - subtract_wt for wt in cum_weights[new_cand_i:]]
         )
     return chosen
-
-
-def apportion(votes: Dict[Constituency, Dict[Candidate, Number]],
-              n_seats: Union[int, Dict[Constituency, int]],
-              apportioner: Union[
-                  'Distributor', Dict[Constituency, int], int, None
-              ] = None,
-              ) -> Dict[Constituency, int]:
-    if isinstance(apportioner, int):
-        # fixed seats for each constituency
-        return {c: apportioner for c in votes.keys()}
-    elif hasattr(apportioner, 'items'):
-        # seats per district are pre-determined statically
-        return apportioner
-    elif hasattr(n_seats, 'items'):
-        # seats per district are determined dynamically outside this
-        return n_seats
-    elif apportioner:
-        constituency_votes = {
-            c: sum(c_votes.values()) for c, c_votes in votes.items()
-        }
-        if isinstance(n_seats, int):
-            # apportion seats across districts by number of votes cast
-            return apportioner.evaluate(constituency_votes, n_seats)
-        elif n_seats is None:
-            # apportionment without fixed total
-            return apportioner.evaluate(constituency_votes)
-        else:
-            raise ValueError('unknown apportionment scheme')
-    elif isinstance(n_seats, int):
-        # delegate to subevaluator
-        return {c: n_seats for c in votes.keys()}
-    else:
-        raise ValueError('invalid apportionment setup')
 
 
 def _select_n_random_float(candidates: List[Any],
